requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
pdf2image==1.17.0
pillow==10.4.0
pyzbar==0.1.9
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
